Create Gardian.master.emerald
Knuckles the Echidna is a fictional character in Sega's Sonic the Hedgehog series. He is a red anthropomorphic short-beaked echidna who is Sonic's best friend and former rival. Determined and serious, but sometimes gullible, he has the ability to glide and climb up walls, and is a powerful fighter with spiked hands.